Add new nav item configuration list, popover and initial drawer. Fix BBUI combobox overflow
