+++ packages/bbui/src/Form/Core/Combobox.svelte
@@ -4,6 +4,7 @@
   import "@spectrum-css/menu/dist/index-vars.css"
   import { createEventDispatcher } from "svelte"
   import clickOutside from "../../Actions/click_outside"
+  import Popover from "../../Popover/Popover.svelte"
 
   export let value = null
   export let id = null
@@ -15,8 +16,10 @@
   export let getOptionValue = option => option
 
   const dispatch = createEventDispatcher()
+
   let open = false
   let focus = false
+  let anchor
 
   const selectOption = value => {
     dispatch("change", value)
@@ -35,11 +38,11 @@
   }
 </script>
 
-<!-- svelte-ignore a11y-click-events-have-key-events -->
 <div
   class="spectrum-InputGroup"
   class:is-focused={open || focus}
   class:is-disabled={disabled}
+  bind:this={anchor}
 >
   <div
     class="spectrum-Textfield spectrum-InputGroup-textfield"
@@ -67,7 +70,7 @@
     tabindex="-1"
     aria-haspopup="true"
     {disabled}
-    on:click={() => (open = true)}
+    on:click={() => (open = !open)}
   >
     <svg
       class="spectrum-Icon spectrum-UIIcon-ChevronDown100 spectrum-Picker-menuIcon spectrum-InputGroup-icon"
@@ -77,41 +80,43 @@
       <use xlink:href="#spectrum-css-icon-Chevron100" />
     </svg>
   </button>
-  {#if open}
-    <div
-      class="spectrum-Popover spectrum-Popover--bottom is-open"
-      use:clickOutside={() => {
-        open = false
-      }}
-    >
-      <ul class="spectrum-Menu" role="listbox">
-        {#if options && Array.isArray(options)}
-          {#each options as option}
-            <li
-              class="spectrum-Menu-item"
-              class:is-selected={getOptionValue(option) === value}
-              role="option"
-              aria-selected="true"
-              tabindex="0"
-              on:click={() => onPick(getOptionValue(option))}
-            >
-              <span class="spectrum-Menu-itemLabel"
-                >{getOptionLabel(option)}</span
-              >
-              <svg
-                class="spectrum-Icon spectrum-UIIcon-Checkmark100 spectrum-Menu-checkmark spectrum-Menu-itemIcon"
-                focusable="false"
-                aria-hidden="true"
-              >
-                <use xlink:href="#spectrum-css-icon-Checkmark100" />
-              </svg>
-            </li>
-          {/each}
-        {/if}
-      </ul>
-    </div>
-  {/if}
 </div>
+
+<!-- svelte-ignore a11y-click-events-have-key-events -->
+<Popover
+  {anchor}
+  {open}
+  align="left"
+  on:close={() => (open = false)}
+  useAnchorWidth
+>
+  <div class="popover-content" use:clickOutside={() => (open = false)}>
+    <ul class="spectrum-Menu" role="listbox">
+      {#if options && Array.isArray(options)}
+        {#each options as option}
+          <li
+            class="spectrum-Menu-item"
+            class:is-selected={getOptionValue(option) === value}
+            role="option"
+            aria-selected="true"
+            tabindex="0"
+            on:click={() => onPick(getOptionValue(option))}
+          >
+            <span class="spectrum-Menu-itemLabel">{getOptionLabel(option)}</span
+            >
+            <svg
+              class="spectrum-Icon spectrum-UIIcon-Checkmark100 spectrum-Menu-checkmark spectrum-Menu-itemIcon"
+              focusable="false"
+              aria-hidden="true"
+            >
+              <use xlink:href="#spectrum-css-icon-Checkmark100" />
+            </svg>
+          </li>
+        {/each}
+      {/if}
+    </ul>
+  </div>
+</Popover>
 
 <style>
   .spectrum-InputGroup {
@@ -124,10 +129,13 @@
   .spectrum-Textfield-input {
     width: 0;
   }
-  .spectrum-Popover {
-    max-height: 240px;
-    width: 100%;
-    z-index: 999;
-    top: 100%;
+
+  /* Popover */
+  .popover-content {
+    display: contents;
+  }
+  .popover-content:not(.auto-width) .spectrum-Menu-itemLabel {
+    width: 0;
+    flex: 1 1 auto;
   }
 </style>
