+++ packages/builder/src/pages/builder/app/[application]/design/[screenId]/[componentId]/_components/Navigation/EditNavItemPopover.svelte
@@ -1,18 +1,32 @@
 <script>
-  import { Icon, Popover, Layout } from "@budibase/bbui"
+  import { Icon, Popover, RadioGroup } from "@budibase/bbui"
   import { createEventDispatcher, getContext } from "svelte"
+  import PropertyControl from "components/design/settings/controls/PropertyControl.svelte"
+  import DrawerBindableInput from "components/common/bindings/DrawerBindableInput.svelte"
+  import DrawerBindableCombobox from "components/common/bindings/DrawerBindableCombobox.svelte"
+  import RoleSelect from "components/common/RoleSelect.svelte"
+  import SubLinksDrawer from "./SubLinksDrawer.svelte"
+  import { screenStore } from "stores/builder"
 
   export let anchor
   export let navItem
+  export let bindings
 
   const draggable = getContext("draggable")
   const dispatch = createEventDispatcher()
+  const typeOptions = [
+    { label: "Inline link", value: "link" },
+    { label: "Open sub links", value: "sublinks" },
+  ]
 
   let popover
   let drawers = []
   let open = false
 
-  $: console.log(anchor)
+  $: urlOptions = $screenStore.screens
+    .map(screen => screen.routing?.route)
+    .filter(x => x != null)
+    .sort()
 
   // Auto hide the component when another item is selected
   $: if (open && $draggable.selected !== navItem.id) {
@@ -25,7 +39,7 @@
     open = true
   }
 
-  const updateNavItem = async (setting, value) => {
+  const update = setting => async value => {
     dispatch("change", {
       ...navItem,
       [setting]: value,
@@ -33,17 +47,7 @@
   }
 </script>
 
-<Icon
-  name="Settings"
-  hoverable
-  size="S"
-  on:click={() => {
-    if (!open) {
-      popover.show()
-      open = true
-    }
-  }}
-/>
+<Icon name={navItem.type === "sublinks" ? "Dropdown" : "Link"} size="S" />
 
 <Popover
   bind:this={popover}
@@ -64,13 +68,64 @@
   maxHeight={600}
   offset={18}
 >
-  <span class="popover-wrap">
-    <Layout noPadding />
-  </span>
+  <div class="settings">
+    <PropertyControl
+      label="Nav item"
+      control={RadioGroup}
+      value={navItem.type}
+      onChange={update("type")}
+      props={{
+        options: typeOptions,
+      }}
+    />
+    <PropertyControl
+      label="Label"
+      control={DrawerBindableInput}
+      value={navItem.text}
+      onChange={update("text")}
+      {bindings}
+      props={{
+        updateOnChange: false,
+      }}
+    />
+    {#if navItem.type === "sublinks"}
+      <PropertyControl
+        label="Sub links"
+        control={SubLinksDrawer}
+        value={navItem.subLinks}
+        onChange={update("subLinks")}
+      />
+    {:else}
+      <PropertyControl
+        label="Link"
+        control={DrawerBindableCombobox}
+        value={navItem.url}
+        onChange={update("url")}
+        {bindings}
+        props={{
+          options: urlOptions,
+          appendBindingsAsOptions: false,
+          placeholder: null,
+        }}
+      />
+    {/if}
+    <PropertyControl
+      label="Access"
+      control={RoleSelect}
+      value={navItem.roleId}
+      onChange={update("roleId")}
+    />
+  </div>
 </Popover>
 
 <style>
-  .popover-wrap {
-    background-color: var(--spectrum-alias-background-color-primary);
+  .settings {
+    background: var(--spectrum-alias-background-color-primary);
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: stretch;
+    gap: 8px;
+    padding: var(--spacing-xl);
   }
 </style>
