+++ packages/builder/src/pages/builder/app/[application]/design/[screenId]/[componentId]/_components/Navigation/index.svelte
@@ -75,7 +75,7 @@
 
   {#if $selectedScreen?.showNavigation}
     <DetailSummary name="Customize" initiallyShow collapsible={false}>
-      <NavItemConfiguration />
+      <NavItemConfiguration {bindings} />
       <div class="settings">
         <PropertyControl
           label="Position"
